+++ griptape/engines/rag/stages/generation_rag_stage.py
@@ -4,7 +4,7 @@ from griptape.engines.rag import RagContext
 from griptape.engines.rag.modules import (
     BaseGenerationRagModule,
     BaseBeforeGenerationRagModule,
-    BaseAfterGenerationRagModule,
+    BaseAfterGenerationRagModule, BaseRagModule,
 )
 from griptape.engines.rag.stages import BaseRagStage
 
@@ -14,6 +14,18 @@ class GenerationRagStage(BaseRagStage):
     before_generator_modules: list[BaseBeforeGenerationRagModule] = field(factory=list)
     generation_module: BaseGenerationRagModule = field()
     after_generator_modules: list[BaseAfterGenerationRagModule] = field(factory=list)
+
+    @property
+    def modules(self) -> list[BaseRagModule]:
+        ms = []
+
+        ms.extend(self.before_generator_modules)
+        ms.extend(self.after_generator_modules)
+
+        if self.generation_module is not None:
+            ms.append(self.generation_module)
+
+        return ms
 
     def run(self, context: RagContext) -> RagContext:
         logging.info(f"GenerationStage: running {len(self.before_generator_modules)} before modules sequentially")