+++ tests/unit/engines/rag/test_rag_engine.py
@@ -23,6 +23,40 @@ class TestRagEngine:
             ),
         )
 
+    def test_module_name_uniqueness(self):
+        vector_store_driver = LocalVectorStoreDriver(embedding_driver=MockEmbeddingDriver())
+
+        with pytest.raises(ValueError):
+            RagEngine(
+                retrieval_stage=RetrievalRagStage(
+                    retrieval_modules=[
+                        VectorStoreRetrievalRagModule(
+                            name="test",
+                            vector_store_driver=vector_store_driver
+                        ),
+                        VectorStoreRetrievalRagModule(
+                            name="test",
+                            vector_store_driver=vector_store_driver
+                        )
+                    ]
+                )
+            )
+
+        assert RagEngine(
+            retrieval_stage=RetrievalRagStage(
+                retrieval_modules=[
+                    VectorStoreRetrievalRagModule(
+                        name="test1",
+                        vector_store_driver=vector_store_driver
+                    ),
+                    VectorStoreRetrievalRagModule(
+                        name="test2",
+                        vector_store_driver=vector_store_driver
+                    )
+                ]
+            )
+        )
+
     def test_process_query(self, engine):
         assert engine.process_query("test").output.value == "mock output"